+++ CHANGELOG.md
@@ -2,6 +2,18 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+## [1.2.0](https://github.com/lightnear/dockerfiles/compare/v1.1.0...v1.2.0) (2022-03-07)
+
+
+### Features
+
+* add chevereto docker image ([097457f](https://github.com/lightnear/dockerfiles/commit/097457f3fac6b3a9372ebfff0179a7fdef013d19))
+
+
+### Bug Fixes
+
+* fix apt-get ca-certificates error ([cf0ff71](https://github.com/lightnear/dockerfiles/commit/cf0ff71e15972337498e53c852d1591f5540cf9e))
+
 ## 1.1.0 (2022-02-23)